Fix vSphere host UUID discovery
When upgrading the vSphere hardware from 5.5 -> 6.5 no more PVCs
can be created. This happens because of a bug on the SEABIOS virtual
hardware that swap the endianness of the content of '/sys/class/dmi/id/pro
on the new virtual hardware versions and now it differs from
'/sys/class/dmi/id/product_serial'.
So, the content in product_uuid is: F5132842-20D4-6171-5DB0-7B96440CCF54
While the product_serial (actually correct and understood by vSphere API)
is:422813f5-d420-7161-5db0-7b96440ccf54

So, we need to use the product_serial instead of product_uuid